Inline components (config, roles, traits)
